test: add E2E tests for DBC filters and update login command
filter-summary-records-cy.ts : Add tests for DBC filters and refactor existing tests to improve test structure and remove duplication of common tasks.
commands.js: Passing password when creating a session in command.js is not necessary; only an identifier needs to be passed to store the session against. This value is output as part of the Cypress run and captured in the reports hence the need to remove the password.
material-selection-list.component.html: add attributes as identifiers for E2E tests to detect