feat: set filter only when valid
In the case where the query param filter is set manually or incorrectly,
the app correctly displays the table according to the default filter
but incorrect filter value is set in state. This value is used to determine
the active state of the filter tabs. If no match found, all tabs are set to
inactive. The highlighted tab is the only indicator of which data view is
currently displayed as there is no view title.