feat:update dbc codes to as environment variables
In order to test that the dbc code changes will work for London DBCs,
they have been added to the environment variables to apply to local
and preprod only.

The switch to a key/pair value is to allow replacing hardcode values in
DBC recommendations filter for env vars.

TIS21-4290: Investigation: Use temporary test Designated Body to
Revalidate in order to build confidence in a smooth transition