refactor:assert that connection page is loaded.
I believe tests were failing due to the login not completing
prior to the test running so that the recommendation page
opened after the visit connections page action.  Hoping the
assertion that waits for the table to display after login
and prior to the tests running will fix.

Also added init connection page function and asserion
that it is the connection page.

NO TICKET