feat: add new reference state and store all DBs
An endpoint to get list of DBCs already exists and
is used by the update connection component to
populate the dropdown list of dbcs and display
connection history. DBCs will now be retrieved upfront
once and added to the store for use where necessary
to avoid unecessary repeated  endpoint calls.

TIS21-4655: Display the DBC as an abbreviation instead
of code in summary records table