feat: add pipe to format designated bodies
Pipe can be used wherever the designated body wants
to be displayed in different format. It can accept any format
as an input and outputs in one of 3 inputs passed as
an argument - either dbc, abbr or name. In case there are
duplicates (unlikely) the pipe will return the original value.

TIS21-4655: Display the DBC as an abbreviation instead
of code in summary records table