feat: remove TIS Designated body column
This column is derived from the programme owner so
displays the same data in different formats. Since the
programme owner is now displayed in same format as
GMC Designated body, the 2 can be easily compared and
TIS Designated body column no longer required.

Lisa also mentioned in the user research session that the
label is misleading since TIS does not have designated
bodies.

TIS21-4655: Display the DBC as an abbreviation instead
of code in summary records table