fix: setting filter state in resolver
Changing tabs in recommendations and connections is opening the
filter panel. This is due to core query params getting added when
switching tabs and so params do then exist. Main change is checking
if table filter params exist. The rest is refactoring to make it
cleaner and more generic.

NO TICKET