fix:move dispatch to resolver
Fetching trainee details data to update state was triggered
in OnInit of details-side-nav component after the details page
was loaded. This resulted in any previous doctor's details held
in state being initially displayed until new data is fetched.

The solution is to move the store dispatch method to the
resolvers for recommendations and connections so that the
data is fetched before the page loads and the details side nav
displays the correct doctor.

TIS21-4868: Prevent details page from initially displaying previous
loaded doctor's details in the trainee profile panel