fix: return both data fetches once complete
Due to the intermittent delay in fetching data from Trainee endpoint
on Prod, the previous fix wasn't sufficient. CombineLatest waits for
items to be emitted from both observables before returning and
resolving to the details page.

The browser won't navigate to the details page until data is returned
so when there is any significant delay, there is no indication in the UI
that the details page is loading. Either the performance issues will
need addressing or some form of loading spinner could be added.

TIS21-4868 - Prevent details page from initially displaying previous
loaded doctor's details in the trainee profile panel