feat: enable sorting on additional columns
Sorting added to GMC no., GMC Submission date and Programme membership.

Programme membership is populated by the field name programmeMembershipType but the ES field is called membershipType, so to enable sorting, we need to add a new property to IRecordDataCell called sortBy which defaults to the name unless specified.

TIS21-5920
TIS21-7938