fix: apply dbc filter in correct location
The adding of DBC to filters from inside the async admin$ pipe didn't make sense and was preventing the dbc filter set in query params from being applied. This in turn was causing the E2E tests to fail as it wasn't filtering by KSS which is the only DBC with a valid RVOfficer role.

NO TICKET