fix: set designated body params correctly
In connections discrepancies, when the tisDesignatedBodies or dbc filter was applied and removed, the programme name filter would always return no results. When set and cleared these filters will be an array containing an empty string, which meant noneSelected was returning false whilst the properties had no value.  These properties can be string or array depending on whether London or not so this has been refactored to test the string length.

Also includes a fix for non-London filter by restoring the valueProperty to the Ready for connection/disconnection table filters previously removed in error.

NO TICKET