fix: display admin name each time rather than email
The template pipe that returns the admin name fails intermitently when the admin data isn't available due to async nature. However, we still want the list to display if the admins endpoint fails or is slow hence adding to the table. In this case it will displaythge default admin email address from ES until the admins are fetched when they are replaced with name via the pipe operator.

TIS21-7779
TIS21-7845